ChatterBot/Gpt2/requirements.txt

git+https://github.com/gunthercox/ChatterBot.git
sqlalchemy<2.1
pytz
spacy>=3.0.0
chatterbot-corpus==1.2.0
PyYAML==3.13
